Moved interfaces to separate directory
