+++ src/LiczbyNaSlowaNetCore/Dictionaries/Currencies/CzkCurrenctDeflation.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using LiczbyNaSlowaNetCore.Interfaces;
 
 namespace LiczbyNaSlowaNETCore.Dictionaries.Currencies
 {
