+++ src/LiczbyNaSlowaNetCore/Interfaces/ICurrencyDictionary.cs
@@ -1,0 +1,18 @@
+﻿// Copyright (c) 2016 Przemek Walkowski
+
+namespace LiczbyNaSlowaNetCore.Interfaces
+{
+    using System.Collections.Generic;
+
+    public interface ICurrencyDictionary
+    {
+        List<string> Unity { get; }
+        List<string> OthersTens { get; }
+        List<string> Tens { get; }
+        List<string> Hundreds { get; }
+        string[,] Endings { get; }
+        List<string> Sign { get; }
+        string[,] Current { get; }
+        bool HasStems { get; }
+    }
+}
